requirements.txt
# Web Framework with known vulnerabilities
Flask==2.0.1          # CVE-2023-30861 (CVSS: 9.8) - Cookie parsing vulnerability
Werkzeug==2.0.0       # CVE-2023-25577 (CVSS: 7.5) - Path traversal

# HTTP Library vulnerabilities
requests==2.25.1      # CVE-2023-32681 (CVSS: 6.1) - Proxy auth leak
urllib3==1.26.4       # CVE-2023-43804 (CVSS: 5.9) - Cookie leak

# Serialization vulnerabilities
PyYAML==5.3.1         # CVE-2020-14343 (CVSS: 9.8) - Arbitrary code execution
                        # NOTE: This version has a known conflict with Jinja2==2.11.3
                        # Upgrading PyYAML requires also upgrading Jinja2 for compatibility

# Cryptography issues
cryptography==3.3.2   # CVE-2023-23931 (CVSS: 6.5) - Cipher weakness

# Template engine
Jinja2==2.11.3        # CVE-2024-22195 (CVSS: 6.1) - XSS vulnerability
                        # NOTE: Older versions conflict with newer PyYAML

# Other utilities
Click==7.1.2
itsdangerous==1.1.0
MarkupSafe==1.1.1

# Testing
pytest>=7.0.0
